Fixes image resizing bug
The previous feature addition broke image resizing because the new attrbiutes
weren't being added to the attrs_order list.  This patch fixes that for this specific case as well as any future cases similar to this.

We do this during the toString function which uses a new helper utility function zencoding#util#fixAttributesOrder which will add any missing attributes in the attr dictionary to the attrs_order list.