Increases configurability of attribute ordering
This change increases the configurability of the attribute ordering by removing
the id and class attributes from being hardcoded in the html lang file.
(It still is in haml and slim as I didn't see that they supported the
default_attributes mechanic).  Instead the id and class are added to a
special key __GLOBAL_ORDER__ in the default_attributes hash in zencoding.vim
This will allow the user to override this should they so desire.

Another option was added to allow adding plain string values to the default
attribute lists for a given tag.  This will allow for sorting attributes that aren't expanded with the tag by default.